+++ packages/live-preview/src/mergeData.ts
@@ -1,10 +1,12 @@
+import type { fieldSchemaToJSON } from 'payload/utilities'
+
 import { traverseFields } from './traverseFields'
 
 export type MergeLiveDataArgs<T> = {
   apiRoute?: string
   depth: number
-  fieldSchema: Record<string, unknown>[]
-  incomingData: T
+  fieldSchema: ReturnType<typeof fieldSchemaToJSON>
+  incomingData: Partial<T>
   initialData: T
   serverURL: string
 }