+++ packages/payload/src/collections/operations/delete.ts
@@ -49,6 +49,7 @@ async function deleteOperation<TSlug extends keyof GeneratedTypes['collections']
       args =
         (await hook({
           args,
+          collection: args.collection.config,
           context: args.req.context,
           operation: 'delete',
         })) || args
@@ -126,6 +127,7 @@ async function deleteOperation<TSlug extends keyof GeneratedTypes['collections']
 
           return hook({
             id,
+            collection: collectionConfig,
             context: req.context,
             req,
           })
@@ -171,10 +173,11 @@ async function deleteOperation<TSlug extends keyof GeneratedTypes['collections']
         // /////////////////////////////////////
 
         result = await afterRead({
+          collection: collectionConfig,
           context: req.context,
           depth,
           doc: result || doc,
-          entityConfig: collectionConfig,
+          global: null,
           overrideAccess,
           req,
           showHiddenFields,
@@ -189,6 +192,7 @@ async function deleteOperation<TSlug extends keyof GeneratedTypes['collections']
 
           result =
             (await hook({
+              collection: collectionConfig,
               context: req.context,
               doc: result || doc,
               req,
@@ -205,6 +209,7 @@ async function deleteOperation<TSlug extends keyof GeneratedTypes['collections']
           result =
             (await hook({
               id,
+              collection: collectionConfig,
               context: req.context,
               doc: result,
               req,
@@ -249,6 +254,7 @@ async function deleteOperation<TSlug extends keyof GeneratedTypes['collections']
 
     result = await buildAfterOperation<GeneratedTypes['collections'][TSlug]>({
       args,
+      collection: collectionConfig,
       operation: 'delete',
       result,
     })