+++ packages/payload/src/fields/hooks/afterRead/promise.ts
@@ -1,6 +1,8 @@
 /* eslint-disable no-param-reassign */
 import type { RichTextAdapter } from '../../../admin/components/forms/field-types/RichText/types'
+import type { SanitizedCollectionConfig } from '../../../collections/config/types'
 import type { PayloadRequest, RequestContext } from '../../../express/types'
+import type { SanitizedGlobalConfig } from '../../../globals/config/types'
 import type { Field, TabAsField } from '../../config/types'
 
 import { fieldAffectsData, tabHasName } from '../../config/types'
@@ -8,6 +10,7 @@ import relationshipPopulationPromise from './relationshipPopulationPromise'
 import { traverseFields } from './traverseFields'
 
 type Args = {
+  collection: SanitizedCollectionConfig | null
   context: RequestContext
   currentDepth: number
   depth: number
@@ -16,6 +19,7 @@ type Args = {
   fieldPromises: Promise<void>[]
   findMany: boolean
   flattenLocales: boolean
+  global: SanitizedGlobalConfig | null
   overrideAccess: boolean
   populationPromises: Promise<void>[]
   req: PayloadRequest
@@ -32,6 +36,7 @@ type Args = {
 // - Populate relationships
 
 export const promise = async ({
+  collection,
   context,
   currentDepth,
   depth,
@@ -40,6 +45,7 @@ export const promise = async ({
   fieldPromises,
   findMany,
   flattenLocales,
+  global,
   overrideAccess,
   populationPromises,
   req,
@@ -193,8 +199,11 @@ export const promise = async ({
           const hookPromises = Object.entries(siblingDoc[field.name]).map(([locale, value]) =>
             (async () => {
               const hookedValue = await currentHook({
+                collection,
                 context,
                 data: doc,
+                field,
+                global,
                 operation: 'read',
                 originalDoc: doc,
                 req,
@@ -211,9 +220,12 @@ export const promise = async ({
           await Promise.all(hookPromises)
         } else {
           const hookedValue = await currentHook({
+            collection,
             context,
             data: doc,
+            field,
             findMany,
+            global,
             operation: 'read',
             originalDoc: doc,
             req,
@@ -266,6 +278,7 @@ export const promise = async ({
       if (typeof siblingDoc[field.name] !== 'object') groupDoc = {}
 
       traverseFields({
+        collection,
         context,
         currentDepth,
         depth,
@@ -274,6 +287,7 @@ export const promise = async ({
         fields: field.fields,
         findMany,
         flattenLocales,
+        global,
         overrideAccess,
         populationPromises,
         req,
@@ -290,6 +304,7 @@ export const promise = async ({
       if (Array.isArray(rows)) {
         rows.forEach((row) => {
           traverseFields({
+            collection,
             context,
             currentDepth,
             depth,
@@ -298,6 +313,7 @@ export const promise = async ({
             fields: field.fields,
             findMany,
             flattenLocales,
+            global,
             overrideAccess,
             populationPromises,
             req,
@@ -310,6 +326,7 @@ export const promise = async ({
           if (Array.isArray(localeRows)) {
             localeRows.forEach((row) => {
               traverseFields({
+                collection,
                 context,
                 currentDepth,
                 depth,
@@ -318,6 +335,7 @@ export const promise = async ({
                 fields: field.fields,
                 findMany,
                 flattenLocales,
+                global,
                 overrideAccess,
                 populationPromises,
                 req,
@@ -342,6 +360,7 @@ export const promise = async ({
 
           if (block) {
             traverseFields({
+              collection,
               context,
               currentDepth,
               depth,
@@ -350,6 +369,7 @@ export const promise = async ({
               fields: block.fields,
               findMany,
               flattenLocales,
+              global,
               overrideAccess,
               populationPromises,
               req,
@@ -366,6 +386,7 @@ export const promise = async ({
 
               if (block) {
                 traverseFields({
+                  collection,
                   context,
                   currentDepth,
                   depth,
@@ -374,6 +395,7 @@ export const promise = async ({
                   fields: block.fields,
                   findMany,
                   flattenLocales,
+                  global,
                   overrideAccess,
                   populationPromises,
                   req,
@@ -394,6 +416,7 @@ export const promise = async ({
     case 'row':
     case 'collapsible': {
       traverseFields({
+        collection,
         context,
         currentDepth,
         depth,
@@ -402,6 +425,7 @@ export const promise = async ({
         fields: field.fields,
         findMany,
         flattenLocales,
+        global,
         overrideAccess,
         populationPromises,
         req,
@@ -420,6 +444,7 @@ export const promise = async ({
       }
 
       await traverseFields({
+        collection,
         context,
         currentDepth,
         depth,
@@ -428,6 +453,7 @@ export const promise = async ({
         fields: field.fields,
         findMany,
         flattenLocales,
+        global,
         overrideAccess,
         populationPromises,
         req,
@@ -440,6 +466,7 @@ export const promise = async ({
 
     case 'tabs': {
       traverseFields({
+        collection,
         context,
         currentDepth,
         depth,
@@ -448,6 +475,7 @@ export const promise = async ({
         fields: field.tabs.map((tab) => ({ ...tab, type: 'tab' })),
         findMany,
         flattenLocales,
+        global,
         overrideAccess,
         populationPromises,
         req,