+++ test/hooks/config.ts
@@ -1,14 +1,17 @@
+import type { SanitizedConfig } from '../../packages/payload/src/config/types'
+
 import { buildConfigWithDefaults } from '../buildConfigWithDefaults'
 import AfterOperation from './collections/AfterOperation'
 import ChainingHooks from './collections/ChainingHooks'
 import ContextHooks from './collections/ContextHooks'
+import { DataHooks } from './collections/Data'
 import Hooks, { hooksSlug } from './collections/Hook'
 import NestedAfterReadHooks from './collections/NestedAfterReadHooks'
 import Relations from './collections/Relations'
 import TransformHooks from './collections/Transform'
 import Users, { seedHooksUsers } from './collections/Users'
-
-export default buildConfigWithDefaults({
+import { DataHooksGlobal } from './globals/Data'
+export const HooksConfig: Promise<SanitizedConfig> = buildConfigWithDefaults({
   collections: [
     AfterOperation,
     ContextHooks,
@@ -18,7 +21,9 @@ export default buildConfigWithDefaults({
     ChainingHooks,
     Relations,
     Users,
+    DataHooks,
   ],
+  globals: [DataHooksGlobal],
   onInit: async (payload) => {
     await seedHooksUsers(payload)
     await payload.create({
@@ -38,3 +43,5 @@ export default buildConfigWithDefaults({
     })
   },
 })
+
+export default HooksConfig