+++ test/live-preview/collections/Pages.ts
@@ -47,7 +47,6 @@ export const Pages: CollectionConfig = {
             {
               name: 'layout',
               type: 'blocks',
-              required: true,
               blocks: [CallToAction, Content, MediaBlock, Archive],
             },
           ],