+++ test/live-preview/config.ts
@@ -1,3 +1,5 @@
+import path from 'path'
+
 import { buildConfigWithDefaults } from '../buildConfigWithDefaults'
 import Categories from './collections/Categories'
 import { Media } from './collections/Media'
@@ -10,6 +12,15 @@ import { seed } from './seed'
 
 export const pagesSlug = 'pages'
 
+export const mobileBreakpoint = {
+  label: 'Mobile',
+  name: 'mobile',
+  width: 375,
+  height: 667,
+}
+
+const mockModulePath = path.resolve(__dirname, './mocks/mockFSModule.js')
+
 export default buildConfigWithDefaults({
   admin: {
     livePreview: {
@@ -19,16 +30,20 @@ export default buildConfigWithDefaults({
         `http://localhost:3001${
           documentInfo?.slug && documentInfo.slug !== 'pages' ? `/${documentInfo.slug}` : ''
         }${data?.slug && data.slug !== 'home' ? `/${data.slug}` : ''}`,
-      breakpoints: [
-        {
-          label: 'Mobile',
-          name: 'mobile',
-          width: 375,
-          height: 667,
-        },
-      ],
+      breakpoints: [mobileBreakpoint],
       collections: ['pages', 'posts'],
+      globals: ['header', 'footer'],
     },
+    webpack: (config) => ({
+      ...config,
+      resolve: {
+        ...config.resolve,
+        alias: {
+          ...config?.resolve?.alias,
+          fs: mockModulePath,
+        },
+      },
+    }),
   },
   cors: ['http://localhost:3001'],
   csrf: ['http://localhost:3001'],